requirement.txt
ipykernel
langchain
python-dotenv
langchain_openai
langchain_core
langchain_community
langchain_text_splitters
bs4
faiss-cpu
pypdf
arxiv
langchain_huggingface
langchain_groq
wikipedia
fastapi
pydantic
langchain-google-genai 
google-generativeai
grpcio==1.67.1 
grpcio-status==1.67.1 
langgraph
langgraph-cli[inmem]
youtube-transcript-api
pytube
streamlit
youtube_dl
accelerate>=0.26.0